Backend/Logic/OCRProcessor/requirements.txt
# Core numerical and image processing
numpy<2
opencv-python==4.8.1.78
Pillow

# OCR
pytesseract

# PDF processing
PyMuPDF==1.22.5
pdf2image

# NLP
nltk==3.9.1

# Document structuring
unstructured[all_docs]

# Optional: pyarrow is excluded to avoid CMake issues on macOS arm64
# pyarrow>=12.0.1
